Update hyperHTML v2.1
The current hyperHTML does not show up in the server list
and it's not configured in an ideal way.

This PR would like to bring latest hyperHTML which is also
oriented to performance rather than correct amount of swaps.

Please consider this merge to make it part of this suite, thanks.