Arrange elements in dropdowns in vertical order instead of horizontal
The framework list is currently arranged like:

A B C
D E F

This patch changes it to:

A C E
B D F

This makes it easier to scan through the list in alphabetical order when
looking for a certain name.

The implementation is done with CSS multi-column layouts. There seems to be
some caveats in the browser support table [1], but this patch only uses the
basic part, so it should be OK.

[1] https://caniuse.com/multicolumn