keyed/svelte: optimize hot loop by eliminating extra whitespace nodes
Svelte does not currently [1] remove purely whitespace text nodes like
most other frameworks do, which causes noticeable performance impact
when benchmarking creating thousands of rows.

This commit manually fixes that so that the generated HTML is closer to
the competing implementations, using a workaround mentioned in [1].

On my system, all the benchmarks involving creation of rows run ~5%
faster with this change and the overall geometric mean improves from
around 1.28 to 1.25.

[1]: https://github.com/sveltejs/svelte/issues/6540