perf, refactor: improve rebuild-all.js
Add or fix JSDoc comments.

Replace deleting files with execSync with fs.rmSync. As a result, the execution time of the rebuildFrameworks function decreased by ~10 times (*1) and memory consumption by 50% (*1).

Improve performance of getting args.

Move the rebuild of the framework into the rebuildFramework function.

Change the framework view from the [dir,name] array to the {type, name} object.

*1 - Measured without installCmd and buildCmd calls.